refactor: move asm packages under internal/
docs/compatibility.md already classified asm, asm/arm64, and asm/amd64 as
"low-level implementation packages that may change without a major version
bump", but nothing enforced it. Moving them under internal/ makes the compiler
enforce the boundary the doc claimed, and stops pkg.go.dev from advertising
~190 exported assembler symbols as consumable API. The module has no release
tags, so there is no source-compatibility cost.

No exported symbol anywhere in the module referenced asm, arm64, or amd64, so
the move is invisible to every public API. cgo, platform build tags, and
abi_arm64.s are file-scoped and survive the move unchanged.

Also convert internal/asm/arm64/abi_arm64_test.go to package arm64_test, the
last non-interp violation of docs/coding-patterns.md §12.2. Its only private
access was a linked.(*caller) assertion; it now calls through the public
asm.Callable interface.

Lost claim, recorded in a doc comment on the test: the concrete-type assertion
kept ctx on the goroutine stack across invoke, exercising a Go-stack-resident
context pointer surviving stack growth and relocation. An interface call
heap-escapes ctx, so that property is no longer covered there. Exporting the
concrete type solely to preserve it would be test-only API widening, which
§12.2 forbids; the interp JIT tests exercise the same trampoline under real
stack growth.

Remove externalize_tests_tmp.go, an orphaned package main at the repo root
from an earlier test-migration sweep, referenced by nothing. golang.org/x/tools
remains a direct dependency via internal/cmd/geninterp.

TestARM64_StackReserve read abi_arm64.s by relative filesystem path, not just
by import; that path is updated too.
