+++ docs/coding-patterns.md
@@ -332,7 +332,7 @@ references.
 | `program` | bytecode aggregates, builders, labels, verification boundary |
 | `interp` | VM execution state, threaded dispatch, fallback, JIT orchestration |
 | `asm` | architecture-neutral assembly IR, allocation, linking, executable buffers |
-| `asm/<arch>` | architecture encoding, ABI, registers, relaxation |
+| `internal/asm/<arch>` | architecture encoding, ABI, registers, relaxation |
 | `prof` | samples, counters, profile aggregation |
 | `pass` | generic pass lifecycle and analysis cache |
 | `analysis` | reusable facts without mutation |
