+++ internal/asm/amd64/arch.go
@@ -3,7 +3,7 @@
 // asm.ErrNotImplemented.
 package amd64
 
-import "github.com/siyul-park/minivm/asm"
+import "github.com/siyul-park/minivm/internal/asm"
 
 type arch struct {
 	registers asm.RegInfo
