+++ internal/asm/assembler.go
