+++ internal/asm/buffer.go
