+++ internal/asm/icache_noop.go
