+++ internal/asm/instr.go
