+++ internal/asm/link.go
