+++ internal/asm/memory.go
