+++ internal/asm/memory_stub.go
