+++ internal/asm/operand.go
