+++ internal/asm/reg.go
