+++ internal/asm/rewriter.go
