+++ interp/cache.go
@@ -5,7 +5,7 @@ import (
 	"sync"
 	"sync/atomic"
 
-	"github.com/siyul-park/minivm/asm"
+	"github.com/siyul-park/minivm/internal/asm"
 	"github.com/siyul-park/minivm/prof"
 	"github.com/siyul-park/minivm/program"
 )
