+++ interp/interp.go
@@ -11,8 +11,8 @@ import (
 	"sync/atomic"
 	"unsafe"
 
-	"github.com/siyul-park/minivm/asm"
 	"github.com/siyul-park/minivm/instr"
+	"github.com/siyul-park/minivm/internal/asm"
 	"github.com/siyul-park/minivm/prof"
 	"github.com/siyul-park/minivm/program"
 	"github.com/siyul-park/minivm/types"
